performanceTests/testy/requirements.txt
argparse==1.2.1
pytest==2.4.2
pytest-timeout==0.3.0
pyhamcrest==1.7.1
requests==1.2.0
requests-futures==0.9.4
arrow==0.3.5
wsgiref==0.1.2
pytest-cov==1.6
pytest-pep8==1.0.4
mock==1.0.1
fabric==1.6.1
babel==1.3
unidecode==0.04.13
docopt==0.6.1
texttable==0.8.1
retrying==1.3.3
shapely==1.5.5
